:sparkles:generate shell autocompletion script
